Remade Sound struct
It now only contains the file name and the audio buffer. This was mostly to ensure that deleting a sound struct is possible and easy, which seemed to be a part of the issue with the previous sound struct. This branch should also be safe to merge into main as of this push